requirements.txt
# Core ML libraries
transformers==4.36.0
torch==2.1.0
datasets==2.14.0
tokenizers==0.15.0
accelerate==0.25.0

# ONNX conversion and optimization
onnx==1.15.0
onnxruntime==1.16.0
optimum==1.16.0

# Data processing and utilities
scikit-learn==1.3.0
numpy==1.24.0
pandas==2.0.0
seqeval==1.2.2

# Synthetic data generation
faker==20.0.0
hebrew-tokenizer==2.3.0

# Configuration and utilities
pyyaml==6.0
tqdm==4.66.0
tensorboard==2.19.0

# Optional: for development and testing
pytest==7.4.0
black==23.11.0
flake8==6.1.0